access restrictions added if not super user 	modified:   app/controllers/bills_controller.rb 	modified:   app/controllers/companies_controller.rb 	modified:   app/controllers/super_users_controller.rb 	modified:   app/views/super_users/index.html.erb 	modified:   app/views/super_users/show.html.erb 	new file:   permissions_temp_file

+++ app/controllers/super_users_controller.rb
@@ -1,48 +1,74 @@
 class SuperUsersController < ApplicationController
 	
 	def index
-		@super_users = SuperUser.all
+		if super_user_signed_in?
+			@super_users = SuperUser.all
+		else
+			redirect_to root_path
+		end
 	end
 
 	def show
-		@super_user = SuperUser.find(params[:id])
+		if super_user_signed_in?
+			@super_user = SuperUser.find(params[:id])
+		else
+			redirect_to root_path
+		end
 	end
 	
 	def edit
-		@super_user = SuperUser.find(params[:id])
+		if super_user_signed_in? && (current_super_user.id.to_s == params[:id])
+			@super_user = SuperUser.find(params[:id])
+		else
+			redirect_to root_path
+		end
 	end
 
 	def update
-		@super_user = SuperUser.find(params[:id])
-		if @super_user.update_attributes(params[:super_user])
-			redirect_to super_users_path
+		if super_user_signed_in? && (current_super_user.id.to_s == params[:id])
+			@super_user = SuperUser.find(params[:id])
+			if @super_user.update_attributes(params[:super_user])
+				redirect_to super_users_path
+			else
+				render 'edit'
+			end
 		else
-			render 'edit'
+			redirect_to root_path
 		end
 	end
 
 	def new
-		@super_user = SuperUser.new
+		if super_user_signed_in?
+			@super_user = SuperUser.new
+		else
+			redirect_to root_path
+		end
 	end
 
 	def create
-		@super_user = SuperUser.new(params[:super_user])
-		if @super_user.save
-			redirect_to super_user_path(@super_user)
+		if super_user_signed_in?
+			@super_user = SuperUser.new(params[:super_user])
+			if @super_user.save
+				redirect_to super_user_path(@super_user)
+			else
+				render 'new'
+			end
 		else
-			render 'new'
+			redirect_to root_path
 		end
 	end
 
 	def destroy
-		@super_user = SuperUser.find(params[:id])
-		if @super_user.destroy
-			redirect_to super_users_path
+		if super_user_signed_in?
+			@super_user = SuperUser.find(params[:id])
+			if @super_user.destroy
+				redirect_to super_users_path
+			else
+				redirect_back_or_to super_users_path
+			end
 		else
-			redirect_back_or_to super_users_path
+			redirect_to root_path
 		end
 	end
 
-	def home
-	end
 end
